Develop to deploy

See merge request s09-final/S09P31A705!230

+++ backend/member-service/src/main/kotlin/com/hotspot/member/controller/MessageController.kt
@@ -60,13 +60,10 @@ class MessageController(
     ): ResponseEntity<MutableList<MessageResDto>> {
         authService.checkAuth(memberId, myId)
         return ResponseEntity<MutableList<MessageResDto>>(
-            messageService.getMyMessageList(myId.toLong()),
+            messageService.getMyMessageList(myId),
             HttpStatus.OK
         )
     }
-
-    // TODO
-    //  일괄 삭제 API 필요
 
     @DeleteMapping("/multi")
     fun deleteMultiMessage(
